Move guard implementation to guard_expr.h/cpp
This is to clearly seperate the two implementations.
The choice between the two will be made in guard.h
